InfoSchema: fix type issues with MySQL 8

+++ src/MYaSyncQL.InfoSchema/Table.cs
@@ -38,7 +38,7 @@ namespace MYaSyncQL.InfoSchema {
         public string TableName { get; set; } = "";
         public string TableType { get; set; } = "";
         public string? Engine { get; set; } = null;
-        public long? Version { get; set; } = null;
+        public int? Version { get; set; } = null;
         public string? RowFormat { get; set; } = null;
         public ulong? TableRows { get; set; } = null;
         public ulong? AvgRowLength { get; set; } = null;
@@ -79,7 +79,7 @@ namespace MYaSyncQL.InfoSchema {
                             newItem.TableName = await rd.GetFieldValueAsync<string>(rd.GetOrdinal(__TableName));
                             newItem.TableType = await rd.GetFieldValueAsync<string>(rd.GetOrdinal(__TableType));
                             newItem.Engine = await rd.IsDBNullAsync(rd.GetOrdinal(__Engine)) ? null : await rd.GetFieldValueAsync<string>(rd.GetOrdinal(__Engine));
-                            newItem.Version = await rd.IsDBNullAsync(rd.GetOrdinal(__Version)) ? default(long?) : await rd.GetFieldValueAsync<long?>(rd.GetOrdinal(__Version));
+                            newItem.Version = await rd.IsDBNullAsync(rd.GetOrdinal(__Version)) ? default(int?) : await rd.GetFieldValueAsync<int?>(rd.GetOrdinal(__Version));
                             newItem.RowFormat = await rd.IsDBNullAsync(rd.GetOrdinal(__RowFormat)) ? null : await rd.GetFieldValueAsync<string>(rd.GetOrdinal(__RowFormat));
                             newItem.TableRows = await rd.IsDBNullAsync(rd.GetOrdinal(__TableRows)) ? default(ulong?) : await rd.GetFieldValueAsync<ulong>(rd.GetOrdinal(__TableRows));
                             newItem.AvgRowLength = await rd.IsDBNullAsync(rd.GetOrdinal(__AvgRowLength)) ? default(ulong?) : await rd.GetFieldValueAsync<ulong>(rd.GetOrdinal(__AvgRowLength));
